Rename functional tests to add FT suffix
