+++ src/functionalTest/java/uk/gov/hmcts/divorce/caseworker/CaseworkerAddNoteFT.java
@@ -12,29 +12,29 @@ import static net.javacrumbs.jsonunit.assertj.JsonAssertions.json;
 import static net.javacrumbs.jsonunit.core.Option.IGNORING_EXTRA_FIELDS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpStatus.OK;
-import static uk.gov.hmcts.divorce.caseworker.event.CaseworkerIssueApplication.CASEWORKER_ISSUE_APPLICATION;
+import static uk.gov.hmcts.divorce.caseworker.event.CaseworkerAddNote.CASEWORKER_ADD_NOTE;
 import static uk.gov.hmcts.divorce.testutil.CaseDataUtil.caseData;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
 
 @SpringBootTest
-public class CaseworkerIssueApplicationTest extends FunctionalTestSuite {
+public class CaseworkerAddNoteFT extends FunctionalTestSuite {
 
-    private static final String REQUEST = "classpath:request/casedata/ccd-callback-caseworker-issue-application-about-to-submit.json";
+    private static final String REQUEST = "classpath:request/casedata/ccd-callback-casedata.json";
 
     @Test
-    public void shouldUpdateCaseDataWhenAboutToSubmitCallbackIsSuccessful() throws Exception {
+    public void shouldUpdateCaseDataWithNotesWhenAboutToSubmitCallbackIsInvoked() throws Exception {
         final Map<String, Object> caseData = caseData(REQUEST);
-        final Response response = triggerCallback(caseData, CASEWORKER_ISSUE_APPLICATION, ABOUT_TO_SUBMIT_URL);
 
+        final Response response = triggerCallback(caseData, CASEWORKER_ADD_NOTE, ABOUT_TO_SUBMIT_URL);
         assertThat(response.getStatusCode()).isEqualTo(OK.value());
 
-        // document_url and document_binary_url are ignored using ${json-unit.ignore}
-        // assertion will fail if the above elements are missing actual value
+        // notes.date value is compared using ${json-unit.any-string}
+        // assertion will fail if the above value is missing
         assertThatJson(response.asString())
             .when(IGNORING_EXTRA_FIELDS)
             .isEqualTo(json(expectedResponse(
-                "classpath:responses/response-caseworker-issue-application-about-to-submit.json"
+                "classpath:responses/response-caseworker-add-notes-about-to-submit.json"
             )));
     }
 }
