--- src/functionalTest/java/uk/gov/hmcts/divorce/citizen/CitizenApplicant1ResubmitTest.java
@@ -1,42 +1,0 @@
-package uk.gov.hmcts.divorce.citizen;
-
-import io.restassured.response.Response;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import uk.gov.hmcts.divorce.testutil.FunctionalTestSuite;
-
-import java.io.IOException;
-import java.util.Map;
-
-import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
-import static net.javacrumbs.jsonunit.assertj.JsonAssertions.json;
-import static net.javacrumbs.jsonunit.core.Option.IGNORING_ARRAY_ORDER;
-import static net.javacrumbs.jsonunit.core.Option.IGNORING_EXTRA_FIELDS;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.http.HttpStatus.OK;
-import static uk.gov.hmcts.divorce.citizen.event.CitizenApplicant1Resubmit.APPLICANT_1_RESUBMIT;
-import static uk.gov.hmcts.divorce.testutil.CaseDataUtil.caseData;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
-import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
-
-@SpringBootTest
-public class CitizenApplicant1ResubmitTest extends FunctionalTestSuite {
-
-    private static final String REQUEST = "classpath:request/casedata/ccd-callback-casedata-applicant1-resubmit-application.json";
-    private static final String RESPONSE = "classpath:responses/response-applicant1-resubmit-application.json";
-
-    @Test
-    public void shouldSendEmailToApplicant1AndApplicant2WhenAllTemplateParamsAreValid() throws IOException {
-        Map<String, Object> request = caseData(REQUEST);
-
-        Response response = triggerCallback(request, APPLICANT_1_RESUBMIT, ABOUT_TO_SUBMIT_URL);
-
-        assertThat(response.getStatusCode()).isEqualTo(OK.value());
-
-        assertThatJson(response.asString())
-            .when(IGNORING_EXTRA_FIELDS)
-            .when(IGNORING_ARRAY_ORDER)
-            .isEqualTo(json(expectedResponse(RESPONSE)));
-    }
-
-}
