+++ src/functionalTest/java/uk/gov/hmcts/divorce/citizen/CitizenSubmitApplicationFT.java
@@ -22,7 +22,7 @@ import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
 
 @SpringBootTest
-public class CitizenSubmitApplicationTest extends FunctionalTestSuite {
+public class CitizenSubmitApplicationFT extends FunctionalTestSuite {
 
     private static final String REQUEST =
         "classpath:request/casedata/ccd-callback-casedata-applicant1-statement-of-truth.json";
