+++ src/functionalTest/java/uk/gov/hmcts/divorce/solicitor/SolicitorSubmitApplicationFT.java
@@ -24,7 +24,7 @@ import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
 
 @SpringBootTest
-public class SolicitorSubmitApplicationTest extends FunctionalTestSuite {
+public class SolicitorSubmitApplicationFT extends FunctionalTestSuite {
 
     private static final String ABOUT_TO_SUBMIT_REQUEST = "classpath:request/casedata/ccd-callback-casedata.json";
     private static final String ABOUT_TO_START_RESPONSE = "classpath:responses/response-solicitor-submit-application.json";
